% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_accoustic.R
\name{acoustic_file_handler}
\alias{acoustic_file_handler}
\title{acoustic_file_handler}
\usage{
acoustic_file_handler(data, fn)
}
\arguments{
\item{fn}{The place to write the plot}
}
\value{
True or False
}
\description{
Internal function. Acoustic data is structured in opposition to project structure so this seperate handler was created to process and write data to the oracle tables
}
