% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_database.R
\name{AddTempRawdata}
\alias{AddTempRawdata}
\title{AddTempRawdata}
\usage{
AddTempRawdata(
  fn = NULL,
  UID = NULL,
  ProjectId = NULL,
  Lat = NULL,
  Lon = NULL,
  Observed_Depth = NULL,
  HaulDate_Start = NULL,
  HaulDate_End = NULL,
  metadx = NULL
)
}
\description{
Internal function. Writes rawdata to database
}
