% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{auto_filter}
\alias{auto_filter}
\title{auto_filter}
\usage{
auto_filter(da = NA, lat = NA, lon = NA, uid = NA)
}
\arguments{
\item{da}{The data to inform}

\item{lat}{the latitude needed to get closest station temperatures}

\item{lon}{the longitude needed to get closest station temperatures}

\item{uid}{the unique id}
}
\value{
dataframe of best guess of down range from da
}
\description{
Internal function. Suggests better bottom times based on various inputs
}
