% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{click.temp}
\alias{click.temp}
\title{click.temp}
\usage{
click.temp(da = NA, af = NA, uid = NA)
}
\arguments{
\item{da}{The data to plot}

\item{af}{The subset of data that is code informed}

\item{uid}{The unique identifier for plotting}
}
\value{
data at new range
}
\description{
Internal function. Further refine the range of when the sensor is at bottom by clicking start and end times on a plot.
}
