% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{decibar2depth}
\alias{decibar2depth}
\title{decibar2depth}
\usage{
decibar2depth()

decibar2depth()
}
\arguments{
\item{P}{pressure, in decibars}

\item{lat}{degrees latitude}

\item{Del}{geopotential anomaly (defaults to 0)}
}
\value{
depths converted from pressure
DEEP-SEA RES., 1976, 23, 109-111.
/ UNITS:
//   PRESSURE     P      DECIBARS
//   LATITUDE     LAT    DEGREES
//   DEPTH        DEPTH  METERS
//   DTN. HEIGHT  DEL    DYN. METERS (geopotenial anomaly in J/kg:: assume 0)
// CHECKVALUE:
//   1.) DEPTH = 9712.653 M for P=10000 DECIBARS, LAT=30 DEG, DEL=0
//   ABOVE FOR STANDARD OCEAN: T=0 DEG CELSIUS; S=35 (PSS-78)
// ----------------------------------------------------------
// Original fortran code is found in:
//   UNESCO technical papers in marine science 44 (1983) -
//   'Algorithms for computation of fundamental properties of seawater'
 \url{http://www.code10.info/index.php?option=com_content&view=article&id=67:calculating-the-depth-from-pressure&catid=54:cat_coding_algorithms_seawater&Itemid=79}, \url{http://www.code10.info/index.php?option=com_content&view=article&id=67:calculating-the-depth-from-pressure&catid=54:cat_coding_algorithms_seawater&Itemid=79}

data from oracle table
}
\description{
Internal function. This function calculates depth in meters from pressure in decibars using Saunders and Fofonoff's method

Provided as an example for presentation
}
