% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{Populate}
\alias{Populate}
\title{Populate}
\usage{
Populate(fn = NA)
}
\arguments{
\item{fn}{The folder path that contains the temperature files to write}
}
\value{
TRUE on success
}
\description{
The prefered method of entry. Writes temperature data to oracle database from files that have formatted header information. See description file for header format information.
}
