% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{Populate_by_worksheet}
\alias{Populate_by_worksheet}
\title{Populate_by_worksheet}
\usage{
Populate_by_worksheet(fn)
}
\arguments{
\item{fn}{The csv file that contains metadata information nessesary for writting temperature data}
}
\value{
TRUE on success
}
\description{
Alternative method of entry. Writes temperature data to oracle database from a worksheet that contains the needed information.
}
