% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_cts.R
\name{SCT_cts}
\alias{SCT_cts}
\title{SCT_cts}
\usage{
SCT_cts(fn, ret, subset)
}
\arguments{
\item{fn}{The data to process}

\item{ret}{List of already procesed SCHEADER data}

\item{subset}{Subset the file by serial number. Cts Data is one big file}
}
\value{
List of formatted data
}
\description{
Function that standardizes cts formatted files
}
