% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_hobo.R
\name{SCT_hobo}
\alias{SCT_hobo}
\title{SCT_hobo}
\usage{
SCT_hobo(fn, ret)
}
\arguments{
\item{fn}{The data to process}

\item{ret}{List of already procesed SCHEADER data}
}
\value{
List of formatted data
}
\description{
Function that standardizes HOBO formatted files
}
