% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_minilog.R
\name{SCT_minilog}
\alias{SCT_minilog}
\title{SCT_minilog}
\usage{
SCT_minilog(fn, ret)
}
\arguments{
\item{fn}{The data to process}

\item{ret}{List of already procesed SCHEADER data}
}
\value{
List of formatted data
}
\description{
Function that standardizes seabird formatted files
}
