% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{standardize.temp}
\alias{standardize.temp}
\title{standardize.temp}
\usage{
standardize.temp(fn, uid = NULL, lat = NULL, lon = NULL, subset = NULL)
}
\arguments{
\item{fn}{The file with sensordata to read}

\item{uid}{the unique identifier if known}

\item{lat}{latitude if known}

\item{lon}{longitude if known}

\item{subset}{subset by Serial_number for CTS data}
}
\description{
Internal function. Convert various data inputs such as minilog and seabird to a common standard used for this project. This method should be build upon for any new data streams
}
