% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperScript.R
\name{trim_to_dates}
\alias{trim_to_dates}
\title{trim_to_dates}
\usage{
trim_to_dates(data, start, end, uid, expand = T)
}
\arguments{
\item{data}{The data to subset}

\item{start}{the start time of desired data}

\item{end}{the end time of desired data}

\item{uid}{the unique id}

\item{expand}{TRUE, want to go beyond start and end if values are the same}
}
\value{
dataframe of new trimed data
}
\description{
Internal function. Simply trim data to specified dates
}
