% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT_database.R
\name{Verified.appendData}
\alias{Verified.appendData}
\title{Verified.appendData}
\usage{
Verified.appendData()
}
\value{
TRUE on success
}
\description{
Update maintables with temp table data. Only do this after manual verrification.
}
